Ported to LLVM3.8, works with GCC 5.4

+++ src/freshvars.cpp
@@ -670,7 +670,7 @@ static void handleStore(Instruction *in, CalledModuleTy *cm, SEXPGuardsChecker *
             return;
           }
       
-          while(freshVars.pstack.size() != newDepth) {
+          while((int)freshVars.pstack.size() != newDepth) {
             unprotectOne(freshVars, msg, refinableInfos, in);
           }
           return;
